pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "manuscript-ocr"
version = "0.1.10"
description = "Detects and recognizes text in historical, archival and handwritten documents. Optimized for fast inference with ONNX Runtime (CPU/GPU/CoreML) and optional training with PyTorch. Supports plugging custom models via a simple API."
readme = "README.md"
requires-python = ">=3.8"
authors = [
    { name = "Pavel Sherstnev", email = "sherstpasha99@gmail.com" },
    { name = "Vlada Zhukovskaya", email = "zhukovskaya.vlada00@gmail.com" },
    { name = "Konstantin Kozhin", email = "i@konstantin-kozhin.ru" },
    { name = "Anna Pyataeva", email = "anna4u.online@gmail.com" },
    { name = "Alina Nurimanova", email = "alenkozavr@gmail.com" }
]
urls = { Homepage = "https://github.com/konstantinkozhin/manuscript-ocr" }

classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Image Processing",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    # ONNX Runtime for inference (CPU by default, GPU/CoreML installed separately)
    "onnxruntime>=1.16",

    # Image processing
    "numpy>=1.21",
    "opencv-python>=4.5.0",
    "Pillow>=8.0.0",

    # Geometry and image processing
    "shapely>=1.8",
    "scikit-image>=0.19",

    # Performance optimization
    "numba>=0.56",

    # Data validation and utilities
    "pydantic>=2.0",
    "tqdm>=4.60",
    "gdown>=4.4",
]

[project.optional-dependencies]

# GPU support for ONNX Runtime (NVIDIA CUDA)
# User installs manually: pip install onnxruntime-gpu
gpu = []

# Apple Silicon support (CoreML acceleration)
# User installs manually: pip install onnxruntime-silicon
silicon = []

# Full development dependencies (includes PyTorch for training)
dev = [
    # PyTorch dependencies for training
    "torch>=1.13",
    "torchvision>=0.14",
    
    # ONNX export tools
    "onnx>=1.12",
    "onnxsim>=0.4",
    "ultralytics>=8.4.30",
    
    # Training utilities
    "torch-optimizer>=0.1",
    "evaluate>=0.4",
    "jiwer>=2.2",
    "matplotlib>=3.3",
    "albumentations>=1.0",
    "tensorboard>=2.4",
    "pandas>=1.3.0",
    
    # Development tools
    "pytest>=7",
    "pytest-cov>=4",
    "black>=22",
    "flake8>=5",
    "isort>=5.10",
    "mypy>=0.991",
    
    # Documentation tools
    "sphinx>=5",
    "sphinx-rtd-theme>=1",
    "numpydoc>=1.5",
    "sphinx-autodoc-typehints>=1.24",
    "sphinxcontrib-mermaid>=0.9",
    
    # Build and publish tools
    "build>=0.8",
    "twine>=4",
]


[tool.setuptools]
include-package-data = true
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]
